backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
